requirements.txt
bleach==6.1.0
cryptography==43.0.1
Flask==3.1.2
flask_apscheduler==1.13.1
Flask_Caching==2.3.1
Flask_Limiter==3.5.0
Flask_Login==0.6.3
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
geopy==2.4.1
numpy==2.4.1
scikit-learn==1.5.2
Pillow==12.1.0
protobuf==6.33.4
python-dotenv==1.2.1
SQLAlchemy==2.0.45
Werkzeug==3.1.5
WTForms==3.2.1
marshmallow==3.21.3
shapely==2.0.6
ortools==9.10.4067
Flask-Babel==4.0.0
Authlib==1.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Flask-Talisman==1.1.0
